fixed jun3

+++ functions/agricola.py
@@ -1042,7 +1042,7 @@ def transform_kissflow_drenajes(drenaje_kissflow):
     df["UBICACIÓN"] = df["UBICACIÓN"].str.upper()
     df["VALVULA"] = df["VALVULA"].fillna(0)
     df["VALVULA"] = df["VALVULA"].replace("",0)
-    df["VALVULA"] = df["VALVULA"].astype(int)
+    df["VALVULA"] = df["VALVULA"].astype(float)
     df["VARIEDAD"] = df["VARIEDAD"].fillna("NO ESPECIFICADO")
     df["VARIEDAD"] = df["VARIEDAD"].str.strip()
     df["VARIEDAD"] = df["VARIEDAD"].str.upper()
